requirements.txt
langchain-text-splitters==0.2.2
openai==1.30.1
qdrant-client==1.17.1
tiktoken==0.7.0
fastapi==0.111.0
gradio==4.44.1
huggingface_hub==0.26.5
uvicorn==0.30.0
python-dotenv==1.0.0
python-docx==1.1.2
pypdf==4.2.0
httpx==0.27.0
